Extraxt Time Helper functions; add new meal option
