+++ src/components/HelperFunctions.ts
@@ -4,27 +4,6 @@ export const getUserId = (user: any): string | undefined => {
   return user?._id || user?.id;
 };
 
-export const formatTime = (hour: number, minute: number): string => {
-  return `${String(hour).padStart(2, "0")}:${String(minute).padStart(2, "0")}`;
-};
-
-export const generateTimeOptions = () => {
-  const options = [];
-  for (let hour = 0; hour < 24; hour++) {
-    for (const minute of [0, 30]) {
-      const hourStr = hour.toString().padStart(2, "0");
-      const minuteStr = minute.toString().padStart(2, "0");
-      options.push({
-        label: `${hourStr}:${minuteStr}`,
-        value: `${hourStr}:${minuteStr}`,
-        hour,
-        minute,
-      });
-    }
-  }
-  return options;
-};
-
 export const sortMealsByTime = (meals: MealData[]): MealData[] => {
   return [...meals].sort(
     (a, b) => a.hour * 60 + a.minute - (b.hour * 60 + b.minute),
