+++ src/views/AddEditMealsView.tsx
@@ -26,15 +26,14 @@ import axios from "axios";
 import { formatISO } from "date-fns";
 import { useNotification } from "../hooks/useNotification";
 import { MealData } from "../interfaces/MealInterfaces";
-import {
-  formatTime,
-  generateTimeOptions,
-  getUserId,
-  sortMealsByTime,
-} from "../components/HelperFunctions";
+import { getUserId, sortMealsByTime } from "../components/HelperFunctions";
 import { ConfirmationDialog } from "../components/common/ConfirmationDialog";
 import { CalendarDatePicker } from "../components/common/CalendarDatePicker";
 import { MealCardItem } from "../components/meal/MealCardItem";
+import {
+  formatTime,
+  generateTimeOptions,
+} from "../components/helperfunctions/TimeHelpers";
 
 const API_BASE_URL = "http://localhost:5000";
